CSS Styling Attempt
Made a .css styling file for the create accounts page, I'm not entirely sure if I did it "right" but I made a css class for all the elements on the page one by one.  Probably redundant but if we want to change anything that spans multiple screens then we can start there.