Sprint 1 req 6 and sprint 2 req 9
Fully implemented across the pages the ability to see your username/profile picture in the top right as well as a logout button replacing the two buttons originally there post login, and the user's name in the top left of the page.

I tried navigating the user to the login page upon logging out but couldn't get it to work.