updated EditJournalEntry.js
fixed some of the issues from before that make the UI of the edit popup to be similar to the Add Journal Entry popup. Added in the logic to deal with journal entry removal from ledger if the edited journal entry was an approved on. Deleted some comments/console.log code in the Journal.js and JournalPage.js.